Revert "HOSTEDCP-1062: Add management cluster KAS network policy"
