+++ control-plane-operator/controllers/hostedcontrolplane/cloud/aws/reconcile.go
@@ -146,7 +146,6 @@ func ccmLabels() map[string]string {
 
 func additionalLabels() map[string]string {
 	return map[string]string{
-		hyperv1.ControlPlaneComponent:       "cloud-controller-manager",
-		config.NeedManagementKASAccessLabel: "true",
+		hyperv1.ControlPlaneComponent: "cloud-controller-manager",
 	}
 }
