+++ control-plane-operator/controllers/hostedcontrolplane/cloud/kubevirt/reconcile.go
@@ -230,7 +230,6 @@ func ccmLabels() map[string]string {
 
 func additionalLabels() map[string]string {
 	return map[string]string{
-		hyperv1.ControlPlaneComponent:       "cloud-controller-manager",
-		config.NeedManagementKASAccessLabel: "true",
+		hyperv1.ControlPlaneComponent: "cloud-controller-manager",
 	}
 }
