+++ control-plane-operator/controllers/hostedcontrolplane/machineapprover/reconcile.go
@@ -206,9 +206,6 @@ func ReconcileMachineApproverDeployment(deployment *appsv1.Deployment, hcp *hype
 			PriorityClass: config.DefaultPriorityClass,
 		},
 		SetDefaultSecurityContext: setDefaultSecurityContext,
-		AdditionalLabels: map[string]string{
-			config.NeedManagementKASAccessLabel: "true",
-		},
 	}
 	if hcp.Annotations[hyperv1.ControlPlanePriorityClass] != "" {
 		deploymentConfig.Scheduling.PriorityClass = hcp.Annotations[hyperv1.ControlPlanePriorityClass]
