+++ control-plane-operator/controllers/hostedcontrolplane/snapshotcontroller/params.go
@@ -46,9 +46,6 @@ func NewParams(
 	params.DeploymentConfig.Scheduling = config.Scheduling{
 		PriorityClass: config.DefaultPriorityClass,
 	}
-	params.DeploymentConfig.AdditionalLabels = map[string]string{
-		config.NeedManagementKASAccessLabel: "true",
-	}
 	params.DeploymentConfig.SetDefaults(hcp, nil, utilpointer.Int(1))
 	params.DeploymentConfig.SetRestartAnnotation(hcp.ObjectMeta)
