+++ control-plane-operator/controllers/hostedcontrolplane/storage/params.go
@@ -39,11 +39,7 @@ func NewParams(
 		ImageReplacer:           ir,
 		APIPort:                 util.APIPort(hcp),
 	}
-	params.DeploymentConfig = config.DeploymentConfig{
-		AdditionalLabels: map[string]string{
-			config.NeedManagementKASAccessLabel: "true",
-		},
-	}
+
 	params.DeploymentConfig.SetDefaultSecurityContext = setDefaultSecurityContext
 	// Run only one replica of the operator
 	params.DeploymentConfig.Scheduling = config.Scheduling{
