+++ test/e2e/util/fixture.go
@@ -144,7 +144,6 @@ func CreateCluster(t *testing.T, ctx context.Context, client crclient.Client, op
 	t.Cleanup(func() { EnsureNoPodsWithTooHighPriority(t, context.Background(), client, hc) })
 	t.Cleanup(func() { NoticePreemptionOrFailedScheduling(t, context.Background(), client, hc) })
 	t.Cleanup(func() { EnsureAllRoutesUseHCPRouter(t, context.Background(), client, hc) })
-	t.Cleanup(func() { EnsureNetworkPolicies(t, context.Background(), client, hc) })
 
 	return hc
 }
